Move handlers definitions to structs package
Because they are structs, after all. Also, in the few following
commits, the daemons structs will be introduced, and they need
to import and use handlers, from which they were spawned. It would
be problematic to import a high-level `reactor.handlers` from
a low-level `structs.containers` (if not moved).
