+++ kopf/reactor/handling.py
@@ -16,7 +16,6 @@ from typing import Optional, Union, Iterable, Collection, Mapping, MutableMappin
 from kopf.engines import logging as logging_engine
 from kopf.engines import sleeping
 from kopf.reactor import causation
-from kopf.reactor import handlers as handlers_
 from kopf.reactor import invocation
 from kopf.reactor import lifecycles
 from kopf.reactor import registries
@@ -24,6 +23,7 @@ from kopf.reactor import states
 from kopf.structs import callbacks
 from kopf.structs import dicts
 from kopf.structs import diffs
+from kopf.structs import handlers as handlers_
 
 WAITING_KEEPALIVE_INTERVAL = 10 * 60
 """ How often to wake up from the long sleep, to show liveness in the logs. """
