+++ kopf/reactor/lifecycles.py
@@ -14,8 +14,8 @@ from typing import Sequence, Any, Optional
 
 from typing_extensions import Protocol
 
-from kopf.reactor import handlers as handlers_
 from kopf.reactor import states
+from kopf.structs import handlers as handlers_
 
 logger = logging.getLogger(__name__)
