+++ tests/authentication/test_authentication.py
@@ -1,9 +1,9 @@
 import pytest
 
 from kopf.reactor.activities import authenticate
-from kopf.reactor.handlers import ActivityHandler, Activity
 from kopf.reactor.registries import OperatorRegistry
 from kopf.structs.credentials import Vault, ConnectionInfo, LoginError
+from kopf.structs.handlers import ActivityHandler, Activity
 
 
 async def test_empty_registry_produces_no_credentials():
