+++ tests/handling/test_cause_logging.py
@@ -4,9 +4,9 @@ import logging
 import pytest
 
 import kopf
-from kopf.reactor.handlers import Reason, HANDLER_REASONS, ALL_REASONS
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.handlers import Reason, HANDLER_REASONS, ALL_REASONS
 
 
 @pytest.mark.parametrize('cause_type', ALL_REASONS)
