+++ tests/handling/test_errors.py
@@ -4,10 +4,10 @@ import logging
 import pytest
 
 import kopf
-from kopf.reactor.handlers import Reason, HANDLER_REASONS
 from kopf.reactor.handling import PermanentError, TemporaryError
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.handlers import Reason, HANDLER_REASONS
 
 
 # The extrahandlers are needed to prevent the cycle ending and status purging.
