+++ tests/handling/test_event_handling.py
@@ -4,9 +4,9 @@ import logging
 import pytest
 
 import kopf
-from kopf.reactor.handlers import ALL_REASONS
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.handlers import ALL_REASONS
 
 
 @pytest.mark.parametrize('cause_type', ALL_REASONS)
