+++ tests/handling/test_no_handlers.py
@@ -4,9 +4,9 @@ import logging
 import pytest
 
 import kopf
-from kopf.reactor.handlers import ResourceChangingHandler, HANDLER_REASONS
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.handlers import ResourceChangingHandler, HANDLER_REASONS
 from kopf.structs.lastseen import LAST_SEEN_ANNOTATION
