+++ tests/handling/test_retrying_limits.py
@@ -5,9 +5,9 @@ import freezegun
 import pytest
 
 import kopf
-from kopf.reactor.handlers import Reason, HANDLER_REASONS
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.handlers import Reason, HANDLER_REASONS
 
 
 # The timeout is hard-coded in conftest.py:handlers().
