+++ tests/hierarchies/test_contextual_owner.py
@@ -4,11 +4,11 @@ import pytest
 
 import kopf
 from kopf.reactor.causation import ResourceChangingCause, ResourceWatchingCause
-from kopf.reactor.handlers import Reason
 from kopf.reactor.handling import cause_var
 from kopf.reactor.invocation import context
 from kopf.structs.bodies import RawBody, RawMeta, RawEvent, Body
 from kopf.structs.containers import ObjectDict
+from kopf.structs.handlers import Reason
 from kopf.structs.patches import Patch
 
 OWNER_API_VERSION = 'owner-api-version'
