+++ tests/invocations/test_callbacks.py
@@ -6,9 +6,9 @@ import pytest
 from asynctest import MagicMock
 
 from kopf.reactor.causation import ResourceChangingCause
-from kopf.reactor.handlers import Reason
 from kopf.reactor.invocation import invoke, is_async_fn
 from kopf.structs.bodies import Body
+from kopf.structs.handlers import Reason
 from kopf.structs.patches import Patch
 
 STACK_TRACE_MARKER = object()
