+++ tests/registries/legacy-2/test_legacy2_resumes_mixed_in.py
@@ -1,7 +1,7 @@
 import pytest
 
 import kopf
-from kopf.reactor.handlers import Reason, HANDLER_REASONS
+from kopf.structs.handlers import Reason, HANDLER_REASONS
 from kopf.structs.resources import Resource
