+++ tests/registries/test_decorators.py
@@ -1,10 +1,10 @@
 import pytest
 
 import kopf
-from kopf.reactor.handlers import ErrorsMode, Activity, Reason, HANDLER_REASONS
 from kopf.reactor.handling import subregistry_var, handler_var
 from kopf.reactor.invocation import context
 from kopf.reactor.registries import OperatorRegistry, ResourceChangingRegistry
+from kopf.structs.handlers import ErrorsMode, Activity, Reason, HANDLER_REASONS
 from kopf.structs.resources import Resource
