+++ tests/registries/test_decorators_deprecated_cooldown.py
@@ -1,7 +1,7 @@
 import pytest
 
 import kopf
-from kopf.reactor.handlers import Activity, Reason, HANDLER_REASONS
+from kopf.structs.handlers import Activity, Reason, HANDLER_REASONS
 from kopf.structs.resources import Resource
